+++ src/Volvox.Helios.Core/Modules/Common/Module.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Discord.WebSocket;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Volvox.Helios.Core.Utilities;
 
@@ -9,16 +10,24 @@ namespace Volvox.Helios.Core.Modules.Common
     /// <summary>
     ///     Unit of the bot.
     /// </summary>
-    public abstract class Module : IModule, IDocumented
+    public abstract class Module : IModule
     {
         /// <summary>
         ///     Unit of the bot.
         /// </summary>
         /// <param name="discordSettings">Settings used to connect to Discord.</param>
-        protected Module(IDiscordSettings discordSettings, ILogger<IModule> logger)
+        /// <param name="logger">Logger.</param>
+        /// <param name="config">Application configuration.</param>
+        protected Module(IDiscordSettings discordSettings, ILogger<IModule> logger, IConfiguration config)
         {
             DiscordSettings = discordSettings;
             Logger = logger;
+
+            var moduleQuery = GetType().Name;
+            Name = config[$"Metadata:{moduleQuery}:Name"];
+            Version = config[$"Metadata:{moduleQuery}:Version"];
+            Description = config[$"Metadata:{moduleQuery}:Description"];
+            ReleaseState = Enum.Parse<ReleaseState>(config[$"Metadata:{moduleQuery}:ReleaseState"]);
         }
 
         /// <summary>
