+++ src/Volvox.Helios.Core/Modules/StreamerRole/StreamerRoleModule.cs
@@ -1,4 +1,3 @@
-using System;
 using System.Linq;
 using System.Threading.Tasks;
 using Discord;
@@ -24,17 +23,11 @@ namespace Volvox.Helios.Core.Modules.StreamerRole
         /// </summary>
         /// <param name="discordSettings">Settings used to connect to Discord.</param>
         /// <param name="logger">Logger.</param>
-        /// <param name="config">Used to access metadata.json</param>
+        /// <param name="config">Application configuration.</param>
         /// <param name="settingsService">Settings service.</param>
-        public StreamerRoleModule(IDiscordSettings discordSettings, ILogger<StreamerRoleModule> logger, IConfiguration config, IModuleSettingsService<StreamerRoleSettings> settingsService) : base(discordSettings, logger)
+        public StreamerRoleModule(IDiscordSettings discordSettings, ILogger<StreamerRoleModule> logger, IConfiguration config, IModuleSettingsService<StreamerRoleSettings> settingsService) : base(discordSettings, logger, config)
         {
             _settingsService = settingsService;
-
-            var moduleQuery = GetType().Name;
-            Name = config[$"Metadata:{moduleQuery}:Name"];
-            Version = config[$"Metadata:{moduleQuery}:Version"];
-            Description = config[$"Metadata:{moduleQuery}:Description"];
-            ReleaseState = Enum.Parse<ReleaseState>(config[$"Metadata:{moduleQuery}:ReleaseState"]);
         }
 
         /// <summary>
